+++ commons/src/main/java/com/simpletools/calculator/commons/operations/base/BitwiseOperation.kt
@@ -2,7 +2,7 @@ package com.simpletools.calculator.commons.operations.base
 
 abstract class BitwiseOperation {
 
-    abstract fun getResult(): Int
+    abstract fun getResult(): Long
 
     abstract fun getBinaryFormula(): String
